SARIi/requirements.txt
Flask==3.1.2
flask-cors==6.0.2
requests==2.32.5
google-generativeai==0.8.6
gTTS==2.5.4
edge-tts==7.2.7
pydantic==2.12.5
beautifulsoup4==4.14.3
googlesearch-python==1.3.0
